refactor theme context
