+++ src/App.js
@@ -1,66 +1,22 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Header from './Header';
 import Content from './Content'
+import themeReducer from './reducers/themeReducer';
+import { createStore } from './createStore';
+import ThemeContext from './themeContext';
 
-// build a store
-const createStore = (reducer) => {
-  let state = null;
-
-  const listeners = []
-
-  const subscribe = (listener) => {
-    listeners.push(listener)
-  }
-
-  const getState = () => state;
-
-  const dispatch = (action) => {
-    state = reducer(state, action);
-    // update the state in the view
-    listeners.forEach((listener) => listener());
-  }
-
-  // init state
-  dispatch({})
-
-  return { getState, dispatch, subscribe }
-}
-
-const themeReducer = (state, action) => {
-  if (!state) {
-    return {
-      themeColor: 'red'
-    }
-  }
-
-  switch (action.type) {
-    case 'CHANGE_COLOR':
-      return { ...state, themeColor: action.themeColor }
-    default:
-      return state
-  }
-}
-
-const store = createStore(themeReducer);
+const themeStore = createStore(themeReducer);
 
 class App extends React.Component {
-  getChildContext() {
-    return { store }
-  }
 
   render() {
     return (
-      <div>
+      <ThemeContext.Provider value={themeStore}>
         <Header />
         <Content />
-      </div>
+      </ThemeContext.Provider>
     )
   }
-}
-
-App.childContextTypes = {
-  store: PropTypes.object,
 }
 
 export default App;
