+++ src/reducers/themeReducer.js
@@ -1,0 +1,16 @@
+const themeReducer = (state, action) => {
+  if (!state) {
+    return {
+      themeColor: 'red'
+    }
+  }
+
+  switch (action.type) {
+    case 'CHANGE_COLOR':
+      return { ...state, themeColor: action.themeColor }
+    default:
+      return state
+  }
+}
+
+export default themeReducer;
