+++ src/themeContext.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import themeReducer from './reducers/themeReducer';
+import { createStore } from './createStore';
+
+const themeStore = createStore(themeReducer);
+
+const ThemeContext = React.createContext(themeStore);
+
+export default ThemeContext;
